Arreglo HU 19

+++ src/main/java/com/dreamteam/arriendatufinca/dtos/PagoDTO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PagoDTO {
-    private Long solicitudId; // ID de la solicitud asociada al pago
+    private Integer solicitudId; // ID de la solicitud asociada al pago
     private String banco; // Nombre del banco
     private String numeroCuenta; // Número de cuenta bancaria
     private Double valor; // Monto a pagar
